requirements.txt
llama-cpp-python
pandas
tqdm
langchain
langchain-community